fix oracle rewrites for other rdbms, add several fixes/features
* add database name in connection msg to syslog
* fix comment commentdata/downtimedata copypaste failures
* fix object insert rewrite for mysql/pgsql
* fix customvariablestatus timestamp

fixes #252,#272

+++ module/idoutils/include/dbhandlers.h
@@ -71,7 +71,7 @@ int ndo2db_handle_commanddefinition(ndo2db_idi *);
 int ndo2db_handle_timeperiodefinition(ndo2db_idi *);
 int ndo2db_handle_contactdefinition(ndo2db_idi *);
 int ndo2db_handle_contactgroupdefinition(ndo2db_idi *);
-int ndo2db_save_custom_variables(ndo2db_idi *,int, int, char *, unsigned long);
+int ndo2db_save_custom_variables(ndo2db_idi *,int, unsigned long, char *, unsigned long);
 
 int ndo2db_log_debug_info(int , int , const char *, ...);
 #endif
